19.2. Service Autowiring Implementation Part 1(Section 3: Modern PHP 7 Introduction (OPTIONAL))
Now it's time to add a new method into the container called Load services which is the old and magic

that we will learn inside this lecture.

And the purpose of this method is to automatically create services for classes from a given namespace

and also how to wire them which means ultimately inject other classes that are also services and are

also out to magically load it using just this one method call.

So we'll be able to get rid of all those service calls and we will get all the services moldered ultimately.

So this is an example how we would call this method load services from our service namespace.

But of course we need to implement all the logic inside load services.

First the metal load services will be able to load the classes and make them services from any given

namespace.

So you can call it many times and it will just load every class inside this namespace and make a service

for it and also automatically inject dependency.

So now let's go to container and start implementing first to actually get a list of classes we need

to find where those classes live.

So let's create a variable and call it base Dira which means base directory will use the DIR constant

which reflects the filename of the current file which is the container.

B and you can see on the right that we based here contains the SLAC folder so it's the full path.

On my system but we've there Esser see folder.

Now as you recall we have passed a namespace an argument for the load services method.

So now we need to convert the namespace to directory structure so we can achieve that by converting

slashes using their SD are Replace function of the standard PDP.

Now you might also remember that our composer configuration was set up so everything that lives inside

a Cerci is ultimately prepended with an app namespace.

You can see on the right namespace is properly converted to the folder structure except that everything

that has up prefix is actually inside.

So c so we need to get rid of that namespace prefix and we could just assume that we have an up prefix

but let's jus for the sake of practice try to remove that prefix programmatically.

So what we are doing here is we are getting the part of the actual directory starting from the first

slash.

Now let's refresh and see how our directory looks right now.

So we got rid of the up prefix but still we have additional slash and we just need to add one to the

SDR pause.

And now you can see that the namespace properly reflects our actual directory structure.

Now you can see the service and controller folders because we have boss to name spaces to the load services

and it was the up service end up controller.

So now we have to actually find the files that are inside the specif it appears specified name spaces

and we can do that by using their Skandia BHB function but we'll find all the files that are inside

specific directory so we are dumping the contents of files and you can see that they are adults and

they're Serializer ph P and index controller ph P.

Now the single dot means the current directory and double dot means directory at one level up.

So we need to filter those from the results we can achieve that using an array filter BHB function and

anonymous function that is a callback to this array filter call.

And what we will do is the first it will call this come dear and it will return whatever you see on

the right which is for each of the namespace says there will be dots blah's the file is found in that

directory.

So we add an anonymous function that will only return true if the file name is different than a single

dot or the double dot to filter out all these dots from her results.

Now you can see we are only left with their actual ph B files.

Now all of this might seem like too much for one lecture so I'll let you have a short break and we'll

see each other again on the next lecture when we will actually use the glass reflection to get some

information about those classes and then we will create services out of Martigny.

So see you in the next lecture.