20.1. Service Autowiring Implementation Part 2(Section 3: Modern PHP 7 Introduction (OPTIONAL))
So now that we have a list of files we can now get some information about those classes and we all use

reflexions for that.

So let's iterate over all of the files that were found.

And for each of those file Let's create reflection class to create a reflection class you need a fully

qualified class name since we don't have the handy Konstanz from the class names we have to use the

namespace and concatenated with the slash and the actual filename

and we'll use the base name.

BHB function for that.

This will just get us the actual file name without that extension.

And we will use the class fully qualified class name that we have it as a service name

so we can use the get name method of the refl. class to get the fully qualified class name.

After you refresh you can see that the class name matches the actual classes class names that we have

and their namespace.

Now we need a way to figure out what are the types of the arguments that are being passed to these classes

which are their dependencies.

So we need a way to figure out what are the types of arguments of constructors of these classes we can

get that by using the reflection class get constructor which will return reflection method instance

which itself has to get arguments method that will return a list of all the arguments of a model.

In our case it does actually get parameters.

So it will return every parameter of constructor.

Now let's create a variable and call it service parameters.

It will hold the names of all the classes class types or interface types that we need to inject into

the constructor of a certain service.

So we'll just figure out what are the names of those parameter types and then we will try to find a

service which has the same name us that type for example the serializer accepts our format interface.

So we will try to find a service which has the same name or Elias.

Now let's see what we can see here.

When we dump this information so the serializer accepts format interface and the index controller accepts

Serializer So it looks correct.

Now it's time to check if their service container already contains a service which has the same name

as dependency.

That has to be passed to the class that we are currently processing.

So we can check for the service name which usually would be a service class or an ally us us.

You might remember we are using aliases to represent interfaces.

So if it contains an alliance or a service we can get service call to the service parameters which as

you can see will just check if the service exists but if it would return no.

Which for example might mean that there is no service for cold format interface will try to fetch it

by Elias.

Now it's the time for the tricky part because if the service is not there it might not necessarily mean

that it's not defined at all.

It might just mean that it wasn't processed yet by our last service says since we are processing file

one by one so it will happen all the time that some service wasn't loaded yet for that will add anonymous

function as service parameter that will actually try to load the service so mind that it is almost exactly

the same as get service and get Elias above.

It's just not fetching it instantly.

It will just call this anonymous function later if the service could not be found yet.

And then it will just try to fetch it later.

So this might look pretty complicated but indeed it is very simple so if the service hasn't been found

because it might not have been loaded yet it lets us load the actual service or fetch the actual service

later when the load services function has time to process all the files that are there.

Finally when we know what are the dependencies of our service we can call the service and the service

name would be the service name variable which is the class name of that particular class we are processing

right now.

And as you remember the second argument is a closure which means an anonymous function.

We will have to use the service name and service parameters because otherwise they won't be accessible

inside the anonymous function because they are out of scope.

Now we need to process the service arguments and since it is all inside anonymous function which will

only be called when the service is being actually fetched from the container.

We can now iterate over all of the service parameters as you see on the top of that code.

They could be either concrete Sarabeth that were fetched or an anonymous function.

So we are checking if it is an anonymous function and the closure is a type for an anonymous function.

We will assign the result of this closure to the service parameter.

Since we are iterating over the service parameters in there for each using the reference operator we

are actually modifying the array so the service parameter is passed to the loop as a reference so we

are just modifying the original element of the array.

And now the same as we defined services manually.

We will return our new class by using the variable.

So we are using the spread operator here and this will just convert the array into set of arguments

to a function.

So using this free dot operator you can convert an array as a list of arguments to a function

now after refreshing the page you can see that suddenly we have two news services and they're called

the same as our classes.

So is up service Serializer end up controller index controller.