20.2. Service Autowiring Implementation Part 2(Section 3: Modern PHP 7 Introduction (OPTIONAL))
So now we can remove the manual servies definitions that we have added and we used some custom names

for them.

The Serializer and controller dot index because they are just out load at.

So now we were testing it using the manual service name controller dot index and instead the new service

name for this controller is actually there up slash controller slash index controller.

Note that you have to use that double slash otherwise it will mean an escape character inside the string.

So now after refreshing the page you can see that everything just works so we have loaded.

Ultimately the index controller and there is an output.

And because their format is on a line us to format the interface it was automatically injected into

the serializer and then into the index controller.

And now let's see how easy it is to pass all these dependencies when we create some more classes because

currently our examples are still very simple.

We have only a couple of classes ban but in a real world projects you'll have hundreds of them.

So it makes the process of configuring them creating services and putting dependencies much simpler

when you have our service container.

So now let's try to output the control or you can see that it also works.

You didn't create this class manually anywhere.

You didn't pass the serializer or even this for math classes and it just Aalto magically work because

we loaded it.

Now you can see that calling both index actions of those two controller classes gives us a result.

It is actually in Dasen because that's why we have configured our service containers so the format servies

which actually had an alliance to form and interface every time form and interface was injected into

one of those services.

It ultimately we have picked up that actually that that Jason class based on service the next lecture

We'll talk about annotations and you will also see how we can implement simple routing which means converting

certain Eurail to the specific controller and it's action.