21.1. Annotations and Kernel Part 1{composer require doctrine/annotations}(Section 3: Modern PHP 7 Introduction (OPTIONAL))
Scenes Annotations are most a part of BHB language.

We need an external library to use them and we can install that on a station's library called doctoring

annotations using the composer require command.

Now let's restart our built in server and get back to work.

Also check out their composer Jason File and how it changed it saved doctoring on a patients library

as a requirement.