21.2. Annotations and Kernel Part 1(Section 3: Modern PHP 7 Introduction (OPTIONAL))
Together with it's version now seems Annotations are just comments that you can put above any methods

or classes.

They are not automatically recognized by Pietsch being because as I said they are not part of the language

itself.

So by adding notations let's add one on top of the post controller but first let's add a new action

by adding a new metaclass called out method.

One and copy the contents of the index and just change the action name so we can see later that it was

actually cold.

So we'll be adding comment on top of the post controller class and this comment will contain a first

notation called root with one parameter which itself is called root.

That will signal our application under which you are well.

This controller will work.

Also we will add another annotation on top of each of the methods of this class and this will be like

part of the road.

So the notation on top of the class will signal what the base root of this controller and on each of

the on top of each of those metals will have specific root so will concatenate the base route together

with the route found above.

Each of those methods.

It will become clearer much later when we actually implement that so don't worry right now.

If you don't really understand what's happening still by adding this notations we've basically added

comments.

So this does not do anything yet.

And that part of the process of using notation annotations is creating a notation class.

Let's start by adding annotations namespace which basically means create annotations directory and inside

it create a road class.

We need a namespace definition so we have our app prefix and the namespace part matching directory which

is annotations.

Create an empty root class and it also needs notation which itself is an annotation.

This is the signal that this class is on on notation and it has one root called one parameter called

root.

Now let's also create a kernel class.

Now kernel as its as its name suggests is a kernel root of our application.

You will also find that Sinfonie has a kernel class Its like a starting point a container of everything

that lives inside the project.

The framework anything so Kerno will also have an instance of our service container.

So we need that private field container to hold the container itself and will define a constructor this

time we are not using that dependency injection because like that's our starting port point of our app

we need to create our first class somehow and the kernel would be our first and only class that we create

explicitly in our index BHB.

It's also useful to have a get container metal that access or metal that will return the instance of

the service container now our kernel needs our boot metal.

That's basically our mettled we're everything is being set up.

I mean that would be definitely that would be our container.

But also anything that we all have in application in the future would be set up in the boot matter.

There will be another one mettled called Boot container from within which we will create will not create

because we create an instance of container inside constructor.