21.3. Annotations and Kernel Part 1(Section 3: Modern PHP 7 Introduction (OPTIONAL))
But we will set up our containers so everything that we did inside index BHB we will do that inside

the kernel class.

So index would be only left with creating new kernel in stands and everything else that you will see

later

so now we will have to add some user statements to the colonel because we are using some classes and

we always need to import all the classes that we use inside certain files.

So for example JSM somehow container format interface.

So we need those statements.

We also do not need Serializer and index controller and probably we won't need a container because they're

both on the same level of the namespace so if class lives in the same directory it verse it has the

same name space or say a day it is on the same level in the name space.

You do not need to use the use statement.

So our boot method will call the boot container.

Now let's go back to our index page index BHB file in which we will create a new instance of the kernel

class and call the boot method to instantiate everything that our application needs.

Currently it is only container.

Now you can instantly see that we can we can fetch the container from a kernel.

So this method is useful.

And that's the only statement we need.

In index it is the up kernel.

Now after refreshing the page you can see we have an error because we copied also the instantiation

code for the container and we actually pass the field inside the boot method.

So this wasn't required anymore.

And that would be all for this lecture.

I understand that some answered questions.

Not everything is clear.

I understand that it is pretty complicated but just place for me to the next lectures where we will

implement everything.

And I hope this will become much more clear how annotations work how we read them how we make something

out of them and what's all this kernel for.

So see you in the next lectures.