22.1. Annotations and Kernel Part 2(Section 3: Modern PHP 7 Introduction (OPTIONAL))
Now let's get Goch working with annotations and open our container class.

So you might have notice that we load the services separately from different name spaces.

And this is for a reason because for controllers we would like to do something extra.

That's why we add an optional closure which would be an anonymous function that can do something or

jus be no.

So it won't do anything.

So if it would be defined we would call it and note that the callback parameter holds a function so

we can call it like a normal beach function and we will pass the service name that was processed and

also the reflection class.

This would be pretty useful for our case because to work with annotations we need reflections now that

doctrine annotations library has it's all Manly's of out-of loading classes to do to remind If the notation

has a corresponding huge B-class that can be loaded.

And since we are using the root notation we need our root class to be loaded so we have to add those

two lines.

The Register load that call from annotation registry.

And then we will need the reader class that will actually read the annotations.

We also need to add and use statement for those two classes and note that they do not have that up prefix.

That's because the library we have installed has its own namespace and its own composer file that defines

an namespace for all its files.

Now you might have remembered that in the last lecture we've added a second optional parameter to the

load services method which could be anonymous function that was called with two parameters the first

parameter was a service name.

And the second one was the reflection class of the service that was currently processed.

So we'll use that now.

And also we need to use the anonymous function to access the variables from the outside scope which

is the roots and the reader instance that instance that we have created.

Now it's time to finally read the annotations from our classes and the reader class has a method called

Get class annotation which accepts the refl. class.

That's why we needed our reflection class and the fully qualified class name of notation that we want

to find.

So as you might see we explicitly tailored for what annotations we are looking for.

Also we have two other use statement for our notation a root class.

Otherwise this would just not work so we have to add this to the place we were where we read it.

And also to our bouffe controllers where we used this notation so that is pretty important.

A use statement for our class.

So using the Get class notation we are just checking if at the class we are working on has annotations

so you can see when you refresh the page that it actually does because we have word on the index controller

and the post controller and they both had their root notation with the slash and the slash posts.