22.2. Annotations and Kernel Part 2(Section 3: Modern PHP 7 Introduction (OPTIONAL))
So if we want find root notation on top of those classes this will read or no.

So in that case we'll just return.

That means we do not have any roots or this class should not have any roots.

Also that there are notations that we create at the root annotation has a root parameter and we will

use that as a base root.

As I explained in the last lecture or so we read this parameter and it becomes the root the root notation

the class parameter.

Now we need to iterate over all of the methods of our controllers which are actions like the index or

the one action and we do the same with the reader.

But this time we call get method and notation to which we need to pass the reflection method instance

which we got from get methods and also the road fully qualified class name which comes from the class

Constand of those class.

Also if we could not find a notation on top of those methods we would continue so we just ignore.

So you can see that we have found free and this makes sense because there is an index and one at that

and also index mettled inside the index controller.