22.3. Annotations and Kernel Part 2(Section 3: Modern PHP 7 Introduction (OPTIONAL))
Now if we are at this point in code that means that we have found a valid route to be that.

So let's add a new entry to the roots array and it will contain the name of the service.

Basically this is the name of the controller service and also a method and the method means the actual

action which is a method of this for example index controller or POS controller we can fetch it using

that reflection metalled get name.

Now let's see what that will create for us by dumping the contents of the roots.

Var. So after refreshing you can see that the first pass has are the index controller service and the

index method.

But the second pass also are that actions from the post controller so we have free index controller

index and post control controller index and post controller 1.