22.4. Annotations and Kernel Part 2(Section 3: Modern PHP 7 Introduction (OPTIONAL))
Now this only is currently plain array.

So we need to add the keys and the keys would be that actually you or else under which we would like

to call this specific service matter.

So we will compose it using the base route which is the annotation on top of each of the controller

classes and the route is the actual route from the method.

So that's the result that we got.

So double slashes slash post and slash posts one.

So that looks valid except those double slashes which we can simply just for now to make this example

working.

Just use SDR replace to replace the instances of double swap slashes with one.

So now we can see on the right that it works and we have our roots.

Now we have our roots but we are not doing anything with them so we know for which you are well what

service and which method of that service should be cold.

But it does not happen here.

So we will add a method on the kernel class called handle request which will basically convert whatever

address that is tied into the browser or whatever client might be used.

And we can get that from the global server variable.

Request your hi.

So let's dump it to see what that gives.

First we all have to add a call to the kernel handle request

and you can see that it's a slash because that's where you are now.

So there is nothing if you will try posts.

It's just slash posts Orris slash posts slash.

So at this point we do not need to call any of them.

Container control or metal container of the controllers from the container directly because we all just

convert this you are now to the specific service that should be called for each of those you or else

so first we need to check if the route that we are on is actually set.

So we will check if the roots field of the kernel which we have to define Of course because it was defined.

If it contains a key we have the current you are I mean this would be relative out to our base server

address.

But first we need to set this roots to the roots variable because we couldn't directly use our class

field inside the anonymous function use statement.

So this just can be done.

So we had to use a separate variable.

So now let's try to find a road using these roots.

You are right.

And assign to a response variable the result of getting a service inside the root we should get to keys

service and method.

So we first tried to fetch the service using container and then using curly braces.

We can call the metalled by not knowing its name.

So the method name is actually being read from a variable and root mettled should contain it.

So well it seems we have an error or

let's find it.

Yes that was the wrong key.

We have other data should be service.

And now you can see that it actually worked.

So for the posts we have called the post and point but let's comment out this on a patient's text.

And also they you or I because we do not need to see this so you can see that for posts we have loaded

the specific action.

Also for the posts one the correct action was being called and also for this address which is index

the correct action from the index controller was called.

So we haven't created any controller explicitly.

We use the service container to load all this and this was done inside the kernel.

And also we use some we use the anonymous function to treat specific services in a specific way.

In this in our case we use that to load the routes for specific services which were controllers also

using this change by returning this from their current so we can immediately call boot and handle requests.

And that's every frame that we have in our indisputably file.

So whenever you liked our new routes you will just need to create a class for them annotate.

Then we have our route on station which is now being read inside the kernel and that's how we can create

very complex applications.

And this is pretty much how the symphony framework is using notations.

It also has a kernel.

It also has a service container.

So after this you should pretty much understand very well how it all will work when we get to the lectures

about Sinfonia and specifically the API platform which relies on patients a lot to configure almost

everything.