52. Validation using regular expressions(Section 7: API Platform - Operations (GET/POST/PUT/DELETE))
You have seen how we verify if the email is correct.

But sometimes you need more advanced validation that you don't have a ready validator for and for that

you can use regular expressions.

So we will the pattern later.

But you can immediately specify on the error message and you will see what are we what we want to achieve

and we want to check if the password is at least seven characters long and whether it contains 1 digit

1 uppercase letter and 1 lowercase letter.

So that is decently secure obviously could be better.

But as it is with their passwords user never remember them.

So let's not make it too complicated and to create our pattern we'll use this reg x hundred one Web

site to visually see how our pattern works on that text.

So first let's maybe write the test string.

So this what we are matching gains against and this would be our Random pass.

Now that dot characters means much any character.

So it has a lot of matches and by using the curly braces we can specify a lens.

So by typing 7 we specify an exact length.

But by adding a coma and not separating their maximum length we match everything that is at least seven

characters so we can see on the right that much information says that the whole string was matched.

Now using parentheses you can add so-called groups.

And the question mark and equals is are so-called positive look ahead.

So this positive Low-Carb will check if something exists in your string before you actually match it

using the dot.

So this is just this will only match that test string.

If those positive lookahead can find something inside the whole string.

So what we did is we marched uppercase characters so these square brackets is a group so we match from

A to Z.

And the case Medhurst So that's why we can add another positive lookahead group.

And this time we are marching.

Any characters that are lowercase and the first the ferde positive lookahead group is that is for the

numbers.

So we will march anything that is zero to nine.

And now we can see that your test string didn't match because we didn't have a number.

So you can have any number of this positive Loughead groups and US I as I said the way they work is

that the whole test string would not match if something from those positive Loughead group could not

be fined found inside the test string.

So if you do not have a number it just doesn't match as you can see on the right on the match information.

Same with lower case and upper case letters and only after the positive lookahead groups we have the

exact matching which is that dot characters and the length of seven or more.

Now we can paste our regular expression expression and mind that you have to surrounded with slashes.

This is how we define regular expressions.

So let's not verify if that worked.

So the user name is too short and we have this is a super secret password.

That now it works still user name is too short but we shouldn't worry about that.

OK and now you can see how we can Vidame more complicated varios like for example passwords.