53. Virtual property (not persisted to database)(Section 7: API Platform - Operations (GET/POST/PUT/DELETE))
It's a common feature on the Internet that when you register on an account you type your password and

you also have to retype it.

So we will the future like dots to our API by adding new field and call it retyped password.

Interesting thing is that we won't store it in the database.

So that's why we do not need a column on notation.

As I said a couple of years ago you only out the column annotation to the fields that you want to actually

store inside the database.

But still you can add the validation constraints because when this data will be sent you can provide

a retyped password and value validated somehow.

But the API platform will only use the fields that are actually inside that actually have the columnar

notation to store them and we'll use the special constraint called expression.

And this you can use to write some simple expressions that are basically Pietsch be and check out this

expression and we are verifying if there this get password which is together for our password will return

exactly the same value as our retyped password and we need to generate together and setter for that

field.

So it is the they are accelerometers as usual so nothing special about that but we need to have them.

So everything works.

Now let's see if our validator actually checks the field that we do not store inside the database.

And you can see that this value is should not be blank and that the passwords do not match.

So that's the message that we have given to this validator this expression validator.

So now let's try to type different password.

And you can see that the passwords does not match or maybe do not match.

And now let's see the successful scenario.

So type the same passwords.

And now the validation has passed.

So as a reminder the validation is being done by device data or component.

But the API platform is calling it behind the scenes after it is serializing the data that we are sending

to it.

It is always also validating it.