54.1. Validating uniqueness of fields (username, email)(Section 7: API Platform - Operations (GET/POST/PUT/DELETE))
By checking the database you can see that we have some behavior that we shouldn't really have.

And this is users having the same usernames and the same passwords.

As I said couple of lectures ago we can verify that.

And this time this would be a constraint at the class level and the name of the constraint is the unique

entity.

So it's also part of the same fine.

It's not implemented by API platform and by adding this unique entity on top of your class when this

class entity is validated it's also checking behind the scenes.

If you fire a record inside the database with the field that you have specified so the default value

for for the unique entity is that the field name that you want to be unique across all the records inside

the table and you can also specify that you want combination of properties to be unique like you want

the user name email combination to be unique.

By specifying it inside the Fields property and surrounding it with curly braces.