54.2. Validating uniqueness of fields (username, email)(Section 7: API Platform - Operations (GET/POST/PUT/DELETE))
Well if you just want the fields to be unique not only when they are together unique so one of them

is unique but you want all of them to be unique no matter what you have to specify two unique entity

constraints and you can see that it works even if we have modified the user name.

So it still throws an error when the email is already exists inside the database.