60. Authentication final configuration and first login using JWT token(Section 8: API Platform - User Authentication and JWT tokens)
So finally after all this configuration and explaining the security configuration we can try to log

in using our username and password and then see if we can get a token which we can then use to authenticate

different requests.

So first let's see this route logging check and try to send username and password to that route.

Mind that unit your server to be started first.

So this is an API logging chain.

The route that we defined as a check BAAF which sinfulness will try to intercept and then authenticate

the user.

So we will choose to send roll inside the postman and then we can write body.

And first you need to said that you want to send application Jason which means that we want the content

type header to be application Jason.

So the API can understand that.

Now we have added our username and password.

So let's try to find a user that might possibly have this username and password may be let's used one

that we have created before or you can find it in the data fixtures.

There are accounts that are created by fixtures and you can see that this route doesn't exist so Symphonie

complaints that it cannot find a route API logging check yet we have defined it as check path and we

said that we don't really have to implement it.

And that Sinfonie intercepts it and then would only then send data to this route.

So yes this is very strange but we don't really have to implement that route.

But anyway we have to define it somewhere inside our application so Sinfonia won't do that for us.

So let's add a new file and call it JPT from inside the routes folder which defines all the routes for

our up and just type what you see on the screen.

So we just fine on BAAF and we do not define any resource to which this should be this request under

that your role should be directed.

Because as I said we won't be doing any actual logic under that you are well but it has to be defined

somewhere so after we have added this route this dummy route actually let's go back to postman and see

if we can authenticate and voila we have a token.

As I said this is handled by jayvee Sukses handler.

You could see that there was a part when it returned and inside the response.

So this is it.

This is out indication talkin that we can now use to authenticate our other requests.

So there are some data encoded inside this token.

This is for example the user name.

That's how it can later figure out what was the user name.

And it is secure only because we are encoding it using our public private keeper.

So each time a user comes back with a token that we have generated we are certain that we have generated

it.

No one could fake it because they didn't have the private or public key.

So also important think never share those public private keys.

And now you can see that our End Point forget user we weren't able to access it.

So we had to change that.

All that API is good all the users could be authenticated and anonymously which means they don't do

not authenticate at all but to actually when we enabled it back sold everything under API requires that

you have authentication talkin of some sorts.

Now we can paste that token that we have generated in our authorization top in postman and choose the

bitter token from the list of possible options.

And when you paste a token that we have generated now we can see that we were successfully able to fetch

our user profile.

So we will be using those tokens for each requests that require authentication.

If you modify the stock and you can see that we have invalid JPT talkin and that's the logic that I

showed you inside the guard authenticator So it is handling cases like that where token is invalid.

It cannot decode it because as you can see we just removed a couple of characters but it makes a big

difference in the public private key cryptography.

So that's it for this lecture in the next sections will handle more topics about security.

But this is that that's it for now.

So I hope this is now much more clear than it was at the beginning.