62. BlogPost validation on POST(Section 9: API Platform - Authorization, Data Validation, Serialization)
Now let's go back to the backbone of our application which is writing blog posts.

Make sure that your server is started or whatever environment you are using.

And let's see if we are able to add on your blog post by sending post requests to their blog posts.

We have their example body that you can find under this lecture so you don't have to write it to yourself.

Note that by selecting certain options in post man headers are automatically added for you.

So the response is loading and we are actually expecting to see an arrow there and that's what we got

because we have disabled their post operation inside the collection operations.

So it's telling us that this operation is like this route is not existing.

And we can change that by modifying their collection operations and now are being bought back the post

which means they're post-operation built in inside their API platform.

If you try again you'll see that we now have an error because we provided all the volleyers but we did

not provide that outer ID.

Since we have modified the blog post entity to also contain Who is there out of that post.

And it should be which should contain the relation to a valid user.

So we'll be fixing that by making this.

You are assigning the currently authenticated user.

But now as you see the outer ID is the last field and we have it set as no because we are not providing

it inside the request and we do not want to do that during the requests because basically when you are

adding a block post it's obvious that it's you adding it.

So you shouldn't be providing the user id or the outer ID.

In this particular case that you you are the outer it should be figure out by the API itself.

So let's start by adding some general validation constraints as you see we have imported the constraints

namespace ASM Asser ally us and we will add some basic constrains to all of those fields to validate

that the data is correct that users do Sandt correct data.

So we are using not Blunk to check whether these properties are sent empty or not.

So we do not want them to be empty.

So we send not blank.

Not that we aren't checking their comments because we won't be doing that right now.

We will also check the length of the content to make sure that those blog posts make any sense.

We would like it to be a minimum of 20 characters.

And also we would like that title to be of at least let's say 10 characters.

So we are doing using the lenth constraint.

We can make it.

So it has a range the minimum or maximum or we can use only the minimum order only the maximum.

Now because every post has to have an out or we will do the same thing as we did for fetching the user

profile and we will access control settings to that post-operation the same way we did.

We've to get Operation get item operation on the user.

So for simplicity we can copy this part from the user anti-Tea make sure that you assign it properly.

So the equal sign and the curly braces and now you said the access control to that function call is

granted.

It is authenticated fully which means that the user has to authenticate first and provide a valid token

for this endpoint to be able to create a new user.

So you can use any of the users from fixtures or just create one because now it works for the users.

So you can just copy their username and password from there another call.

And now let's try to authenticate so we can have valid Dockum that we will have to provide to the block

posts posts the post request.

We do it the same as usual.

Choose to Beera talkin and bastes are valid.

Talkin to that request.

If you press the preview request in the authorization tap you can see how the authorization header looks

like so it contains that Beyrle string and the contents of our token.

Now we have added the validation constraints so they are immediately making validation for us checking

for all the things that we have defined for example that the contents should be at least 20 characters

or more.

But anyway this brings us to an error because we still do not provide the outer ID so this is a fatal

error.

Basically it comes from the database that outer ID cannot be know and we are not providing it in any

way and we will fix that by reading inside the API.

What's the ID of the current user and assigning it to every new block post that is being created.